move run api test to test_cli.py
