+++ tests/autorag/test_cli.py
@@ -3,6 +3,10 @@ import pathlib
 import subprocess
 import tempfile
 
+from click.testing import CliRunner
+
+from autorag.cli import cli
+
 root_dir = pathlib.PurePath(os.path.dirname(os.path.realpath(__file__))).parent
 resource_dir = os.path.join(root_dir, 'resources')
 
@@ -23,3 +27,11 @@ def test_evaluator_cli():
         assert os.path.exists(os.path.join(project_dir, '0', 'retrieve_node_line'))
         assert os.path.exists(os.path.join(project_dir, '0', 'retrieve_node_line', 'retrieval'))
         assert os.path.exists(os.path.join(project_dir, '0', 'retrieve_node_line', 'retrieval', '0.parquet'))
+
+
+def test_run_api():
+    runner = CliRunner()
+    result = runner.invoke(cli, ['run_api', '--config_path', 'test/path/test.yaml',
+                                 '--host', '0.0.0.0', '--port', '8080'])
+    assert result.exit_code == 1  # it will occur error because I run this test with a wrong yaml path.
+    # But it means that the command is working well. If not, it will occur exit_code 2.
