+++ tests/autorag/test_deploy.py
@@ -5,10 +5,8 @@ import tempfile
 import pandas as pd
 import pytest
 import yaml
-from click.testing import CliRunner
 from fastapi.testclient import TestClient
 
-from autorag.cli import cli
 from autorag.deploy import summary_df_to_yaml, extract_best_config, Runner, extract_node_line_names, \
     extract_node_strategy
 from autorag.evaluator import Evaluator
@@ -190,11 +188,3 @@ def test_runner_api_server(evaluator):
     assert len(retrieved_contents) == 10
     assert isinstance(retrieved_contents, list)
     assert isinstance(retrieved_contents[0], str)
-
-
-def test_run_api():
-    runner = CliRunner()
-    result = runner.invoke(cli, ['run_api', '--config_path', 'test/path/test.yaml',
-                                 '--host', '0.0.0.0', '--port', '8080'])
-    assert result.exit_code == 1  # it will occur error because I run this test with a wrong yaml path.
-    # But it means that the command is working well. If not, it will occur exit_code 2.
